Added solidity prettier for code formatting for contractsSol5, main contracts (batch 1) (#854)
* Added prettier-solidity package

* Prettified .sol files files, main contracts

* Add EOF empty line .prettierrc

* Fixed unwanted formatting in KyberNetwork.sol

+++ contractsSol5/IBurnableToken.sol
@@ -1,5 +1,6 @@
 pragma solidity 0.5.11;
 
+
 interface IBurnableToken {
-    function burn(uint _value) external returns(bool);
+    function burn(uint256 _value) external returns (bool);
 }
