+++ contractsSol5/IGasHelper.sol
@@ -4,6 +4,12 @@ import "./IERC20.sol";
 
 
 interface IGasHelper {
-    function freeGas(address platformWallet, IERC20 src, IERC20 dest, uint tradeWei, 
-        bytes32[] calldata t2eReserveIds, bytes32[] calldata e2tReserveIds) external;
+    function freeGas(
+        address platformWallet,
+        IERC20 src,
+        IERC20 dest,
+        uint256 tradeWei,
+        bytes32[] calldata t2eReserveIds,
+        bytes32[] calldata e2tReserveIds
+    ) external;
 }
